Add temp (?) fix for events on clusters
Some of our canvas events code was not playing nice with cluster nodes,
so this introduces some quick fixes. We may or may not keep these
changes, as they relate to cluster nodes and edges not being added to
the `nodes` and `edges` objects (they are tracked internally by the
canvas).