Fix multiple cluster bug
This was caused by the discrepancy between our `edges` object and the
list of edges the vis.js canvas tracks on its own. We don't currently
add cluster edges to our list, but vis.js does. We may want to
re-evaluate this in the future, but for now it suffices to say that it
might be overkill to try to track clustered edges ourselves. It's a lot
easier to just check if an edge is "clustered", a "cluster edge", or a
"normal edge".