Add new cluster grouping rule
From now on, a user can only cluster connected nodes if each of those
nodes' neighbors is connected to each other candidate cluster node. This
is meant to prevent the scenario where a neighbor node is associated
with a cluster, but it isn't associated with EVERY node in the cluster.