Fix stale data bug in context menu
Forgot to clear the selected node edge labels object when the user
selected more than one node, or any edge. This wouldn't currently cause
any problems in the view, since we don't use that input when either of
the described cases are true, but this could have caused bugs in the
future. Better to clean up stale data than to let it persist.