Fix sizing issue for cluster nodes
Note: it was a mistake to remove the 'size' option from the initial
canvas config. While it is true that certain shapes of nodes are not
affected by the 'size' option, cluster nodes (and perhaps other
entities) are! So by removing the size option, cluster nodes size was
none, which is why they appeared to be missing.