Add duplication of nodes when clustering
This came up during a review meeting. It might be confusing for users if
the edges they didn't cluster on disappeared after clustering.
Therefore, the duplication of nodes to prevent this from happening. I
don't think this is really how vis.js is intended to be used, so we're
being a bit hacky with this approach. We will have to be very careful
with keeping the data consistent between our components and vis.js's
internal model.