Add view fixes leftover from merge
Adds a bit of a hacky solution for the issue of not being able to see
the visualization search bar because of the mat-sidenav container: just
push down the absolutely-positioned sidenav container. Not the best
solution in the long-term -- and I've created a ticket explaining as
such -- but for now it works fine.