Clean up duplicate node/cluster bar chart code
Lots of changes here! Next step is to actually test all these features
so we can be more confident that this all is working the way we intend
it to. Going to create a PR for these changes as there doesn't seem to
be anything breaking that hasn't been explicitly identified in comments
or otherwise.