Add settings.json and linting doc
Also removed .vscode and the settings.json from the list of ignored
VSCode files. I am not planning on fixing all the linting issues present
in the app in this branch, so expect many errors until that is done.

I also added a small developer doc describing how/when to lint, and some
QOL tips for doing linting.