Add direction check to clustering algorithm
Came to the realization that we need to check direction of edges when
clustering. This actually hasn't been a problem up until this point, but
eventually we will run into the edge case of a node having two edges,
where they point in opposite directions but have the same edge label.