Add post-merge cleanup
Also adds some changes to the drawing tool spec testing module. I was
going to fix the tests but I now realize that the problems may be very
tightly coupled to the design of the drawing tool component. It may be
difficult for me to fix the issues without doing some redesign.