New annotations that includes text positions.
Refactored the TokenExtractor into AnnotationsPDFParser. The class now parses PDF taking into
account the text positions. The refactored sequential increment function extract_tokens() now
returns a data structure called PDFTokenPositionsList, which contains the individual keyword
tokens, as well as the index of each character within the keyword. This index will be used to
map to the LTChar objects that contains the coordinate positions of the character.