Add clipboard service
We should be able to use this service anywhere we want to read/write to
the user's clipboard. I imagine we will want to read more often than we
want to write, since it will probably be better to create a separate
service if we want to cache certain data.