Fixed issue for detecting of parts of a keyword is on a new line.
Previously the algorithm was detecting when a part of a keyword is on a new line incorrectly. Added
a ratio of the height to the calculation. Additionally, added some refactor to clean up code,
as well as a new synonym typo correction. Still TODO: PDFs that have no spacing at all - both
synthetic and actual spacing. These PDFs are still causing an issue on how to handle them.