Refactored token generation algorithm to work for PDFs with no space.
It is still not perfect with PDFs that have no space. The reason is because they insert
virtual characters like \n char. Keeping these in the token position objects that are
generated can cause, and removing them can also cause problems. For now, remove them
and turn them into whitespaces. The small occasions in which this causes a problem is
when a word continues onto the next line. By changing it into a whitespace, the
sequential increment algorithm will see each as separate words. Oddly, this doesn't happen
in all occurences, will need to look again...