fix: problems with pathnames on /files/<id>
Not sure about the source of the issue, but maybe the folder
for output files was not existant, and this would generate
an exception at the time of write.

Other problems with the current implementation:
- the function is not thread-safe, in the sense that
  if two users try fetching the same pdf, the function
  would try writing to the same file at the same time,
  potentially returning malformed data or generating other
  unexpected issues
- written files are not cleaned up, try using temp files

The real solution here would be to avoid writing to the filesystem.
It should not be necessary.
Attempt to fix all this at: https://github.com/biosustain/text-mining-backend/pull/1
I will work more on it asap.

chore: remove unused (global) variables