Add interval tree to find all annotations that have overlapping intervals.
For now do nothing with these conflicting annotations. JIRA LL-407 and LL-408 was originally
suppose to be handling these conflicts, but we decided to hold off until we get a better
idea of what we want to do with them. So just leave them there for now as a place to gather
all conflicts.