Refactor Part II
Decided to go with removing the environmental URL setting for services
as this didn't make sense. If we used relative URLs, our webserver
(e.g.) nginx proxy will perform the redirects of domain/api calls for
us. The proxy.conf.json in Angular is a substitute for this when
developing locally.