refactor: reannotate expects a list of ids instead of a single id
feat: enable multi-select in File Browser's table

Selecting more than one row disables the "Open" button
Selecting zero rows disables both the "Open" and the "Reannotate" button

feat: add an annotation_status property to the PdfFile interface

This is used to add feedback in the File Browser table.
An interface AnnotationStatus is also added, to specify valid statuses.
This should go away once we have a way to assess an annotation age.
It is assumed that, when we fetch the list of files available, these are all
successfully annotated.

feat: File Browser's table shows the annotation status (check mark, cross mark or ellipsis)