Fixed parsing and coordinate bugs.
The extract_token was not considering part of a keyword that's on a new line.
The _create_keyword_objects function needed to return the coordinates of the parts
of a keyword that's on a new line. Because we need to consider the upper right
coordinates to see if it's larger than the upper right coordinates of the
part that's on the previous line.