Add initial gene/organism matching fix
This commit fixes two of the cases described in LL-358: Case 1 and Case
3. The effect of this (assuming we have NCBI data loaded into the graph
database), is that all the fluff genes we were seeing before will be
eliminated from the document. If there is no NCBI data in he graph
database, then the effect will be that NO genes are annotated.