refactor: /reannotate returns more info
It informs the client whether the annotation:
- succeeded
- failed
- the file was not found
for each of the files.

A response for a request of annotating three files could look like the following:
{
	"file_id_1": "Annotated",
	"file_id_2": "Not annotated",
	"file_id_3": "Not found"
}

Outcome values are defined by an Enumeraton named AnnotationOutcome

refactor: Implement interface Reannotation

This is to reflect changes in the backend, described above.

chore: add debug logs in masterToggle()