Add handler for 3rd case in gene/organism matching
Adds a check for the most frequently occurring organism match for each
gene. In the case that a gene name appears in more than one organism, we
use the organism which appears most frequently in the document. We may
fine-tune this heuristic later. For instance we could calculate the
distance between the gene and its nearest matched organisms, and then
take the closest one.