Add temp lookup table for popular organisms
This is a temproary alternative to looking up in Robin's Neo4j data. We
currently have no reliable way to identify which strain a gene belongs
to, so instead we use a lookup table for a handful of popular organism
strains and use the genes there. This should still enable us to filter
out many crummy genes, just not as many as before. We should re-evaluate
these tradeoffs in the future.