Update entity location function to add human ids
The homo sapiens entity ID is added to the list of organism locations at
any virus location in the document. This is because we assume that when
genes are mentioned near viruses, the author is referencing human genes.
It's possible that this is not always the case, and while unlikely, it
is also possible that homo sapiens and a virus might have a gene with
exactly the same name. In this case, we should always default to the
human gene. This check still needs to be added.