Add service to get legend from the backend
This is going to need some tweaking, because there's a few issues:
first, the type labels don't have a secondary color, so we might want to
add that. Second, the labels are lowercased, which means they won't be
recognized by Neo4j. We may consider have a second dictionary which is
specifically for the visualizer, but keep it on the backend. In any
case, this ticket will need refinement.