Added a call to ord() to check character unicode.
There are cases like an ellipse that will not work when we remove trailing punctuation.
The reason is because it is a single char, but when we remove trailing, that will
capture only one of the periods, and the index for that is removed in the dict.
This causes an index out of bounds when moving onto the next period.