Fix expandOrCollapseNode behavior
Also removes a few data structures which were no longer necessary. We
were keeping track of duplicate edges which had been created via
clustering so we could avoid drawing those edges. The reasoning for this
was driven by the assumption that there should never be more than one
copy of an edge drawn on the canvas, either on the map or abstracted
into a cluster. This is no longer the case. Both the real and the
clustered version of an node/edge can appear on the map now.