Add fix to clustering behavior
Previously, a node would always be either completely put into a cluster
or disconnected from the hub node, depending on what other nodes (if
any) it was connected to. Now, if a node is connected to a hub, it stays
connected when clustered EXCEPT if its only connection is to the hub
node.

In addition, the "Group by Rel" option is a bit smarter now: it won't
include relationships where all the connected nodes are already in
existing clusters connected to the hub.