Add cancel effect when node is selected
Previously the edge/cluster snippets requests would interrupt each
other, but would not be interrupted by a node being selected. This
commit fixes this issue, and also fixes an issue where the sidenav was
not immediately swapped over to the new entity type when one was
clicked. Before the sidenav would update only when the request finished,
which is obviously not ideal.