Got dupsort to work.
The py-lmdb docs were not very clear. Apparently dupsort only works if the database
handle was not the main database. Had to add max_dbs and create a second database in the .mdb file.
Also fixed the kibana indexing script to work correctly with parallel_bulk and switched to
using a generator instead of storing in memory with a list. This was causing a sigkill for out of memory
due to genes and eventually pubchem having 30+ million entries.