Add misc quick fixes to viz search
Includes several temporary solutions to miscellaneous issues in the
visualizer search. Ideally, these issues would be solved for both the
viz search and for the drawing tool search, and both tools would use the
same search service. However for now we've opted to use a quick fix for
the visualizer.