Some human genes were annotated as rat genes.
This is a side effect of using the gene common name in the query to KG. EDEM3 vs Edem3 result
in two different gene annotation having the same interval. We cannot differentiate and choose one
randomly. This fix proposes to lowercase the results from the KG because at that point, we already
have the genes with their gene names case sensitive.