Integrate NLP into annotation pipeline.
Had to refactor the AnnotationsPDFParser class a bit due to the NLP service returning index
that do not match. The reason is when putting together words for the sequential walk,
we do not include double spaces - but do keep the correct index. However, the old way
of combining the words together into a large string mean the NLP service will return
slightly different index because it didn't know there was a double space.

Also added a check for admin role to only show annotation option if admin.

Also minor update to the NLP service and blueprint code.