Refactored the pdfminer parsing to not include double spacing.
Previously, the code that checks whether or not to add a virtual space was checking if the char was a space char,
but it could also be a newline or a newline space. So changed it to check for those as well. Now the index mapping
matches what comes back from the NLP service. Also cleaned up the NLP service code, removing unused code. NLP service
continues to crash with no errors and exit code of 247...