style: Improve node-search-bar code before a refactor
I plan to remove the filtering and fetch logic from it, as per the single responsibility principle.
This became necessary once I had to add another filter to this part of the web app (the filter for organisms).
I think the logic should get out of this component, and stay in the parent component, instead.

Ideally, the bar simply outputs the string that the user enters. The "filter" components outputs all info about filters. And the "list" component simply lists the result that THE PARENT component fetches.

Summary:

Removed the form.
Removed two non-unique id attributes from the page (probably the result of a copy-paste from other components:
- search-bar-container in mat-card
- search-bar-full-width in mat-form-field
Removed attribute mat-button from the button.
Use 2-spaces indentation.

It'd be nice to get rid of the mat-card, too. It's not visible in the UI. Doing so produces artifacts that I don't wanna deal with, right now.

Use absolute paths in imports.
Remove unimplemented OnInit hook.
onSubmit body is more readable.