Update gene import queries to be a bit faster
Removes the col & row properties on imported nodes, and adds batching of
the queries. An important takeaway from some of the analysis that was
done on the queries was that the more properties the user adds on the
input nodes, the slower the query is going to be. This may be helped by
adding an index to the imported nodes. Perhaps we can create a import
specific node label and generate an index on that.