Fix viz cntxt menus not working in workspace tool
The original assumption that there would be exactly on context menu on a
given page wasn't going to work in the workspace tool. The only thing
that needed to change (that we've caught so far) is the element ids
being non-unique with more than one instance of the context-menu. Now,
all tooltips generated from the tooltip component will create a unique
id upon construction, but the developer still needs to utilize the id.