Fix import data slice table w/ large strings
We needed to handle the scenario where any of the table cells had a very
long string. The default behavior was to extend the table to fit the
content of the cells, which wasn't very user-friendly. This could still
use some improvement, it would be nice if the headers could be different
sizes but still fit to content.