Refactored files.py blueprint a bit.
For now, extracted out some queries into models/files_queries.py and refactored them to be better performant.
Previous implementation queried the database multiple times which can be slow if the list of files
were extremely large. Instead do it in one query and do updates and deletes in bulk on a lower level for faster performance.