Fixed center point calculation when determining whether an annotation is custom annotation.
Since a word term can span multiple lines, need to check the center point of each individual rectangle.
Previously was calculating the rectangle incorrectly, and resulted in a skewed rectangle for multi-lines.
The custom annotation uses the coordinates from the backend, we can compare the two and check the
center point for each rectangle in the list.

Also refactored the if statement for skipping LTAnno to avoid out of index error, and added a try/except
when creating keyword object.