Changed when the stop words are taken into consideration.
Previously was using the stop words when making the tokens, but this is not right. So reverted back to the
previous common word set for that part. Instead use the stop words when doing LMDB lookups for all except species.
Also refactored and fixed a bug that appeared due to saving the file before annotating. Now, files are saved
permanently even if annotations fail. This will allow the user to view the document even if annotations
fail, and gives them a chance to reannotate or reupload.