Replace fk to files with fk to file_content
This replaces the foreign key on global list to files with a foreign key
to file_content. The reason for this was because of a bug Binh pointed
out where we could not delete a project file without either deleting any
global inclusions/exclusions linked to that file, or nulling the foreign
keys.

This change also slightly refactors the global list endpoints to use a
new raw pdf file download api. They previously had been using a project
file endpoint, but that will no longer work with the new foreign key.