Fixed PDF parsing and keyerror for LL-1497.
The PDF parsing had an edge case where the space was not being added, which resulted in a
concatentation of two words, so a potentially annotation was not being created. Also added
so if statements to avoid key error when doing gene to organism matching. This occurs when the
user creates a custom annotation that is a number. We currently do not create word tokens that
are numbers.