Fixed bug in re-annotating with custom species annotation.
Previous implementation was flawed because it was passing the id of an entity even when the
entity did not match with the data in custom annotations, this resulted in a KeyError.
Additionally, the data created from the custom annotation was also flawed and should
have been structured similar to LMDB. Also removed the extra call to Neo4j to get
custom species taxonomy id as that was not needed because the id was already present
in self.organism_frequency.