Implemented global inclusions for all entity type.
Global inclusion for Gene, and potentially Protein in the future, need to be reworked. The reason is
the inclusion would not be in the KG for Genes, otherwise it would have been annotated. The correct
approach would be to use the gene_id and query the KG to get the main gene name. Then use that gene
name to do gene organism matching. Similar process for Protein.