Refactored entity lookups.
Previously, to lookup an entity, LMDB would be used just to validate the entity exists, then at a later
time LMDB is used again to actually create the annotations. This is inefficient because of having
to process LMDB and do the lookup multiple times. Now, instead of validating, do the lookup at that
step in addition to checking the global inclusions. Performance is faster and equals the
performance before logging was added.