Refactor elastic file search class/implementation
Previous document search was pdf specific, which doesn't make sense now
that we've refactored the index to use both pdf and maps. Refactors the
old elastic search query code in the ElasticService, and also updates
the 'pdf-search' endpoint to filter out any maps/pdfs the user doesn't
have access to.