Fixed slow interval tree.
Before when creating the interval tree, the annotations object were added as data in the AnnotationInterval.
But this caused the tree to be extremely large and take a long time to create. Instead, only
use the intervals to create the tree. Then use a dict lookup to get the annotations for that interval.
This is faster, and timing it took .15 seconds for 10000 nodes without the data, but took 31 sec with data.