Fix resizing issues in word cloud
Fixed the issue by enforcing stricter sizing on the word cloud toolbar
buttons. Actual issue seemed to be that the word cloud and its parent
div were getting out of sync with each other because of the resizing of
the toolbar buttons, so this should prevent the issue we were seeing
before. A user might see this happen again if they do some weird
vertical/horizontal resizing, but it shouldn't happen unless they go out
of their way to make it happen.

Also cleaned up the spacing on the toolbar elements in both the
word cloud and the pdf viewer.