Add highlight + doi to DirectoryObject interface
Needed to add two new optional attributes to the interface, otherwise
angular would complain on production build. These attributes are only
present for DirectoryObject when querying elastic, and are only used in
the content search. Also fixes a bug where we were showing an empty list
of search terms in the file browser preview for pdfs.