Optimized global inclusion and exclusion.
For global exclusion, for every LMDB lookup, an exclusion set was created for each entity.
This became a bottleneck when the global exclusion list is long, so create the sets beforehand.
Additionally, for global inclusion of genes, group the gene_ids together to make one call to
the KG, rather than make one for every inclusion.