Move default filtering to word cloud init
Was doing the filtering by having the annotation filter emit to the
parent, but this is not so great because it means the cloud has to be
drawn twice. Ideally we ought to have a word cloud component that
doesn't do any filtering, but right now it's sort of assumed that we
will always have it. May sub-class it out later.