Changed the data class for tokens to be more straight forward.
Instead of using dicts to keep track of individual character positions, just keep a list.
Because the characters are concatenated in order, the coordinates, height, width etc
will also be in order when added to the list. So the index will map directly to each
character. This reduces the complexity to get individual index, and is much cleaner.

Doing it this way also makes the jsonb dump of the combined words much smaller.
Previously was saving each character as a json block, which was inefficient,
but now save each word (pre sequentialed combined). These words also hold
data needed for checking abbreviations, cropbox, etc. Which also saves having
to pass around a bunch of data.

TODO: Update AnnotationsService, unit tests, etc to use these new classes.