Add greek symbols and removed clean_char function call from normalize_str.
Removed clean_char because we only want to clean specific ones, and keep things like greek symbols.
If a token starts with delta, strip out the symbol but keep the index. This will allow us to
match on the gene name, and also still have the coordinate index of the symbol to annotate later.