Added executemany_mode and executemany_values_page_size.
The psycopg2 driver for postgresql is not effective by default. There are helper functions that can
be used to improve the performances. By default, the driver will execute a statement for each paramenter
in the sequence passed to executemany, but the helper functions will fix that and group them
into one query separated by commas instead.