Fixed id hyperlink and prefix.
Previous implemented was overlaying the custom annotations over the systems annotation.
This did not work if the inclusions were global, as another user uploading a different paper
would not have the hyperlink and db source type from the original user. So the fix was to keep
track of the id_type and id_hyperlink. If empty, these will default to what we've been doing,
but if not empty, they will use what is there.