Refactored code structure and inheritance.
Create two new DBConnection and GraphConnection class. These will be the base of postgres services
and graph services. Previously we had (still do - need to refactor the rest) these two classes in
one class, but now it should be easier to do separation of concern. Moving this way will help us
refactor code easier in the future.

Also added migration to add primaryName to custom annotations, and re-update system annotations
to query the knowledge graph for anatomy and food primary names.

Custom (or manual) annotations will add primaryName when it is first created.