Removed expensive _get_global_inclusion_values function.
Function was called ~1.6 million times, totaling up to cummulative time of 28 secs. Removed it and used
inline lookup instead, resulted in shaved off that much time in annotating. Also moved the matched_type
to outside of the lookup functions to avoid adding to the call stack when it was already matched.