Revert advanced args usable in query string
Because we need to search projects exactly, we're forced to use the ID
as the search term. It seemed odd to me to put the ID of the project in
the query string, and that realization snowballed into just removing the
non word/phrase tokens from the query string entirely...It remains to be
seen whether this change will itself be reverted, but I think this is
the better option for now.