Add limit to cloud algo input
Previously we had no logical limit on the entity cloud visualization
algorithm, which meant we were often doing a lot more work than was
necessary to produce the cloud.

Solution was to simply add a hard limit on the number of words given as
input to the algorithm. The idea is that a user's viewport is only ever
going to be so big, so all we need to do is choose a reasonable upper
limit on the number of words that could be in the cloud.