Annotation pipeline integration with NLP.
User can choose how to annotate an entity type, use NLP or Rules Based. For entities selected for NLP,
the pipeline will call the NLP service. The results from the NLP will acts as a veto not a suggestion.
So any results from NLP will be cross-referenced and if NLP says no, then that token is not that entity.

The token will still go through all LMDB, the veto will act as a stop gap and skip that LMDB.