Remove whitespace validator from query param
Having this whitespace validator felt strange to me from a UX
perspective. If the user wants to put in a weird input, that's fine as
long as it's at least a string. Alternatively, we can simply trim
anything they give us without them ever knowing it. The end result is
that instead of getting an error, they simply get no results.