Refactored enrichment table annotations.
Previously, each cell was being annotated individually and this caused a significant slowdown on performance.
The reason was due to constantly setting up and tearing down the connection to neo4j, and running the annotation
pipeline from start to finish for each cell. The refactor combined all cells into one text string, and annotated
like that. Additionally, to avoid removal of overlaps, the annotations are split prior to the interval tree, and
re-combined after. Performance is still a problem, but is now mostly a frontend issue to draw the highlights.

Final performance for ~300 genes, so 300 rows and ~10 columns each is about 27-30 seconds.