Fine tuned error handling.
Changed Exception class to ServerException with a default user displaying message. Consolidated some exception
classes, and changed the pattern of non user-friendly error handling. For the non user-friendly errors,
use the pattern of logging then raising. By doing this, the raising will default to the default
message, and the logging will have the message for devs to look at. Fixed the interface for errors on
the frontend so it will actually display the message.