Add find feature to pathway prototypes
There are a few caveats with this feature, and I woujldn't currently
consider it production ready. First and foremost, we need to upgrade our
vis.js version. It seems there are a few issues that are fixed by more
recent versions. Second, while this feature works, it requires the
entire network to be redrawn whenever a user enters a new search term.
Clearly far from ideal, and I'm not sure there's a good solution to the
issue.