Add update whenever annoData changes in wc filter
Recently added a change that allowed us to stop using an ugly timeout to
know when to draw the word cloud. However, this introduced a side effect
where the annotation filter could be loaded before the annotation data
had arrived, which led to the bug described in LL-2655. Fix was to
always update the annotation filter data whenever annotationData came
in, not just when the component is first loaded.