Fix leftover issue from original fix
Originally, I had updated the refresh of annotationData to occur anytime
the input was set to a new value. However, I failed to realize this
meant not only when the parent changed the input value, but also if we
set a new value inside the annotation filter component, which we did
when the annotations were sorted.

The fix was to update this portion of code to sort the array in place,
rather than create a sorted copy and assign it as the new value.