Updated gene/protein organism matching algorithm.
Previously, the closest organism was either above or before. If a fallback was given, and the distance
of the closest was greater than 200, the fallback was used if it matched, if there was no match then
it defaulted back to the closest organism.

The new matching algorithm adds an extra step if the distance is greater than 200, and either no
fallback was given or fallback did not match. At this point, find the closest organism before, if
non found then do not annotate.