Removed extra API call to the KG to create the enrichment JSON.
The JSON is already saved in file_content table, so no need to recreate it. Also changed the edit parameters to
save instead of refreshing. This will allow the file to update the file_content. The downside is the user
will now have to manually click refresh data. Also moved all the enrichment text mapping work to the backend.