Refactored how enrichment table snippet tags are created.
Using regex is slow and not accurate. So use the index of the word instead. To do this, had to fix a minor
bug regarding keyword length. Previously was taking the length of the text in document, but this is not
consistent because the PDF parser could potentially strip out punctuation that are actually part of a word.
So with the new implementation of keyword length, use that to determine the start/end index and create
the snippet tags from there. This improved snippet tag creation to be 100 seconds faster.