Performance refactor for annotation pipeline.
Reverted interval tree for fixing annotation overlaps to not use pandas because it is slower. Also opt for in-line code when generating sequential tokens in EntityRecognition because it is faster.
Iterating through dict.items() is slow, so changed the enrichment mapping to be a list and result is significant improvement in iteration time.