Readded optimized enrichment KG queries.
Also reverted back to the optimized data processing steps. But also modified it a bit because
a gene can point to two different synonyms, and vice versa for synonyms to genes. This updated
processing also fixes JIRA LL-2989 where the column data is completely off and not related to the genes.