Fixed minor bug in generating sequential tokens. Essentially at the start of the sequence, do not ignore any terms that fit removal criteria.
We will remove them at the very end. The reason is because a real term could start with one of those words. If they
were removed too earlier, there is a small chance of having a duplicate with another word.

E.g 'I am foo' could become 'am foo' which will be a duplicate of 'am foo' when 'am' is the start of the sequence.