Refactored how LMDB is implemented to hopefully avoid issues with unexpected environment behaviors during multiprocessing.
The class LMDBAccess is now LMDBConnector and implements the __enter__ and __exit__ to take advantage of using it
inside a with block statement. The code was also refactored to avoid overloading the class with multiple class
variables for each env and transaction. Now there is a single open_db function to handle that.