Update elastic search query to return less data
Previously we were returning the entire document, in addition to the
highlight fragments elastic found within the document. This was
producing responses in the tens of MB, which was a huge waste. Now, we
only return the ID of the doc, and the highlights, which results in a
huge decrease in the response size (in general).