Major refactor of annotation related service classes.
Previous implementation resulted in a tight coupling of several classes, this made unit testing difficult
because the EntityRecognition was calling external databases for data, and mutating the instance variables.

The refactor separated out those database calls to the AnnotationDBService and AnnotationGraphService. These
are called first, and passed into the __init__ for EntityRecognition. This should be cleaner and make it easier
to write unit tests.