Added need_review property when inserting global inclusions.
We only insert if main node exist and synonym does not or main node does not exist, this ensures only newly
create global inclusions will execute the create query. So only they will have the need_review, this will help
differentiate between existing nodes in the KG when we need to review and remove bad inclusions.