Ignore unnecessary props when expanding nodes
Still getting all properties at first, because there is no way of
knowing which properties we need until we examine what the types of the
expanded nodes are. For example, the "display name" of a node might be any one
of: "description", "name", "title", or "displayName". Since we don't
know what properties the expanded nodes will have, we filter out
unnecessary properties later.