Replace FileNodeData with FileSchema
FileNodeData was originally intended to allow us to return data from the
hierarchy endpoint with exactly what we needed. However, because
FileSchema was designed to interface with the FilesystemObject class,
and that class has many helper methods, I decided to use FileSchema
instead.

What this means is we'll have more consistency, at the cost of returning
more data than we actually need. There may still be a way to reduce the
data returned, but it would likely involve extending some of
Marshmallows behavior.