Revert change to `update_projects` function
Had originally changed this because I thought the current code would not
trigger the ORM event handlers, but it seems I was mistaken (or perhaps
SQLAlchemy was updated and this issue no longer occurs). Either way the
change I made to handle this is not necessary, so I've reverted it.