Drop the organism_gene_match table.
(1) Fixed migration 782b6ba8b7eb to be able to migrate with the OrganismGeneMatch model deleted.
(2) Also fixed a minor bug that was suppose to prioritize gene common name, but it was prioritizing synonyms.

Things to keep in mind, with the multiple data sources for genes, we are now losing some genes, e.g CRP.
This gene comes back with two data sources, both have the same synonyms and common name. The dict only keeps
the last entry, so we need to either change the usage of dict or have a data source precedence.