Changed the key used for data_sources lookup and moved the check prior to checking if synonym is in the dict.
The reason for the change in key is to check at the synonym level, but also consider the organism id. This
should be unique enough to handle the cases where the synonyms are the same, but the gene names are different.
As well as when both gene name and gene synonym are the same. E.g CRP has two rows with gene name/synonym being the same.
Then there is FRA2 which has the synonyms being the same, but the gene name being FRA2 and BOL2. For the FRA2 case,
we want to go with the ncbi BOL2 one, so having the if check prior to the synonym check should fix this.