Changed logic implementation to gene/organism matching to better account for different combinations of genes/organism.
Since we've always only chosen the first gene/organism, we can use this logic in record keeping. This guarantees we will only
ever have one gene/organism and we can choose to keep the first NCBI that we find. Doing it this way, we can move back
the choosing of first gene/organism to AnnotationGraphService, and no longer need to do next(iter(...)) in the AnnotationService.