Refactor content search query parser
Introduces `pyparsing` as the primary parsing tool for the content
search. Previously we were using regular expressions, but with the
introduction of more sophisticated boolean expressions into the search
query, this was going to quickly become messy. `pyparsing` allows us to
easily parse the input query and directly translate it into an elastic
query.