Remove score from kg search return value
Including the score is what was causing duplicate results. A node would
appear twice in the results if it had two synonyms for the entered term,
and the synonyms had different fulltext match scores. We don't actually
use the score, and the fulltext results are automatically sorted, so
removing score was the easiest solution.