Add fix for LL-3406
Logic for generating the hierarchy object was a little bit off. I think
this was due to some bad copy pasta. Essentially the bug was caused by
the most recent parent dumping all of its children. I.e., if a parent
folder was encountered after its children, all the children would be
removed from the tree.