Fixed some issues regarding export and global creations.
The query failed due to the type not being in the dict, so just switched the if statement around. Add a check for file id since
it is possible we are using the same graph for different environments.

Also fixed the create global queries to use a :GlobalInclusion label per Robin's suggestion to make it possible to query for an
existing real data source synonym node.